Snap packaging
This commit introduces snap packaging for osmclient.

Snaps are a fancy zip file containing an application and all of its
dependencies, and a description of how it should run safely on a
user's system. This is a multi-distribution alternative to installing from source, or debian package.

Snaps are designed to be secure, sandboxed, containerised applications isolated from the underlying system and from other applications. Snaps allow the safe installation of apps from any vendor on mission critical devices and desktops. Updates are done atomically, rolling back automatically in the case of failure, with support for release channel and series.

More info on snaps can be found here: https://snapcraft.io/

Signed-off-by: Adam Israel <adam.israel@canonical.com>