bug 397 Preserve env running install-openmano.sh
This is a partial bug fix for bug 397, by making sure that sudo
is not invoked inside the install-openmano.sh script (which would
discard the environment already set, at least on default Ubuntu 16.04).
It is safe to not invoke sudo because the script requires to be run as
root. Additionally, it is now recommended to use "sudo -E" when invoking
the script, instead of just "sudo", to make sure the environment is
preserved.
The "sudo -E" recommendation has also been added to
install-openmano-service.sh.

Change-Id: Ib8fc5807e26f7018d9e5331c8b0c045794346d5a
Signed-off-by: Igor D.C <igor.duarte.cardoso@intel.com>