Add unittest to exercise python compatibility
It was recently discovered that the IM module is not Python3 compatible
due to issues with the pyangbind library. This patch adds a specific
test that will exercise basic parsing of a VNFD, which passes under
Python2 but fails under Python3. This will be useful for verifying
Python3 support when it lands in pyangbind.

Signed-off-by: Adam Israel <adam.israel@canonical.com>