Pin the libjuju version
There is a pending release of libjuju that will break compatibility with
the version we're currently using. This won't be a problem when we
switch to using the version within N2VC, but there's more work involved
with that.

This patch just installs the specific version we've been using, 0.6.1,
so that new installs won't get the new version.

Signed-off-by: Adam Israel <adam.israel@canonical.com>