Fix itermittent error installing Juju
We originally made the decision to disable apt update and upgrades for
faster machine provisioning, but we've experienced the scenario where
the provisioning of a container will fail because the apt cache in the
lxd image is stale. This can cause the cloud-init used by Juju to get
stuck in a loop trying to install packages that no longer exist in the
Ubuntu Archive.

This patch enables the run of `apt-get update` when Juju provisions a
container. This will add some time to the provisioning process, but
ensures greater reliability in the process.

Signed-off-by: Adam Israel <adam.israel@canonical.com>