Fix Docker build
This fixes the docker build, that was failing while installing N2VC due
to a new version of websockets that was released.

We specifically install the latest 4.x release of websockets, known to
be compatible with libjuju, in conjunction with a patch to N2VC that
specifies the version requirement.

Installing the specific version is necessary because websockets is being
installed as a dependency to another python library that is installing
5.0, but installing 4.1 satisfies its requirements.

Change-Id: I390b2c3663f21b29e8527783d153127ca405ea8e
Signed-off-by: Adam Israel <adam.israel@canonical.com>